merge

+++ nhom7APPDEV/database/migrations/2023_03_17_062633_create_trainee_courses_table.php
@@ -15,15 +15,14 @@ return new class extends Migration
     {
         Schema::create('trainee_courses', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('user_id');
- 
+            $table->unsignedInteger('user_id');
+
             $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedBigInteger('course_id');
- 
+
+            $table->unsignedInteger('course_id');
+
             $table->foreign('course_id')->references('id')->on('courses');
             $table->timestamps();
-        
-           
         });
     }
